Update source to pass type-checker.

+++ src/UI/ProseMirror/EditRange.js
@@ -1,56 +1,176 @@
 // @flow strict
 
 import type { Transaction } from "prosemirror-state"
-import type { Mark, Node, Schema } from "prosemirror-model"
-import { Decoration, DecorationSet } from "prosemirror-view"
-import { getSelectionMarkers } from "./Marks"
+import type { Mark, Node, Schema, NodeRange } from "prosemirror-model"
+import { getSelectionMarkers, getMarkersAt, type Marker } from "./Marks"
 import { Selection } from "prosemirror-state"
 import { Fragment, Slice } from "prosemirror-model"
 import Parser from "../Allusion/Parser"
 import Serializer from "../Allusion/Serializer"
+import { Decoration, DecorationSet } from "prosemirror-view"
+import type { Mapping } from "prosemirror-transform"
 
-export class Range {
-  +index: number
-  +length: number
-  constructor(index: number, length: number) {
+export interface Range {
+  index: number;
+  length: number;
+}
+
+export class EditRange implements Range {
+  index: number
+  length: number
+  decorations: DecorationSet
+  constructor(index: number, length: number, decorations: DecorationSet) {
     this.index = index
     this.length = length
+    this.decorations = decorations
   }
-  static fromNodeRange({ start, end }: { start: number, end: number }) {
-    return new Range(start, end - start)
+  static empty = new EditRange(Infinity, 0, DecorationSet.empty)
+  static attributes = { nodeName: "span", class: "edit" }
+  static options = {
+    inclusiveStart: true,
+    inclusiveEnd: true
   }
-  static fromSelection({ from, to }: { from: number, to: number }) {
-    return new Range(from, to - from)
+  static new(index: number, length: number, doc: Node) {
+    if (length === 0) {
+      return EditRange.empty
+    } else {
+      const decoration = Decoration.node(
+        index,
+        index + length,
+        EditRange.attributes,
+        EditRange.options
+      )
+      return new EditRange(
+        index,
+        length,
+        DecorationSet.create(doc, [decoration])
+      )
+    }
   }
-}
+  static fromNodeRange(range: NodeRange) {
+    return EditRange.new(range.start, range.end - range.start, range.$from.doc)
+  }
+  static fromSelection(selection: Selection) {
+    return EditRange.new(
+      selection.from,
+      selection.to - selection.from,
+      selection.$anchor.doc
+    )
+  }
 
-export const editableRange = (selection: Selection): Range => {
-  const { $cursor, node } = selection
-
-  if ($cursor) {
-    const node = $cursor.node()
-    switch (node.type.name) {
-      case "heading":
-      case "anchor": {
-        return Range.fromNodeRange($cursor.blockRange())
+  map(mapping: Mapping, doc: Node) {
+    if (this.length === 0) {
+      return this
+    } else {
+      const decorations = this.decorations.map(mapping, doc)
+      const decoration = decorations.find()[0]
+      if (!decoration) {
+        return EditRange.empty
+      } else {
+        const { from, to } = decoration
+        const length = to - from
+        if (this.index !== from || this.length !== length) {
+          return new EditRange(from, length, decorations)
+        } else {
+          return this
+        }
       }
     }
   }
-
-  if (node) {
-    switch (node.type.name) {
-      case "image":
-      case "horizontal_rule": {
-        return Range.fromSelection(selection)
-      }
+  patch({ index, length }: Range, doc: Node) {
+    if (length === 0) {
+      return EditRange.empty
+    } else if (this.index === index && this.length === length) {
+      return this
+    } else {
+      return EditRange.new(index, length, doc)
     }
   }
-
-  return markedRange(selection)
+  includesSelection({ from, to }: Selection): boolean {
+    const { index, length } = this
+    return from >= index && index + length >= to
+  }
+  includes({ index, length }: Range): boolean {
+    if (length === 0) {
+      return true
+    } else if (this.length === 0) {
+      return false
+    } else {
+      return index >= this.index && this.index + this.length >= index + length
+    }
+  }
 }
 
-export const markedRange = (selection: Selection): Range => {
-  const markers = getSelectionMarkers(selection)
+export const editableRange = (selection: Selection): EditRange => {
+  const { $cursor, node, $anchor } = selection
+  const { doc } = $anchor
+
+  if (!$cursor) {
+    return EditRange.empty
+  } else {
+    let depth = $cursor.depth
+    const { nodes } = $cursor.parent.type.schema
+    while (depth > 0) {
+      const node = $cursor.node(depth)
+      switch (node.type) {
+        case nodes.paragraph:
+        case nodes.heading: {
+          return EditRange.new(
+            $cursor.start(depth) - 1,
+            $cursor.end(depth) + 1,
+            doc
+          )
+        }
+      }
+      depth--
+    }
+    return EditRange.empty
+  }
+
+  // if ($cursor) {
+  //   const node = $cursor.node()
+  //   switch (node.type.name) {
+  //     case "heading":
+  //     case "anchor": {
+  //       return EditRange.fromNodeRange($cursor.blockRange())
+  //     }
+  //   }
+  // }
+
+  // if (node) {
+  //   switch (node.type.name) {
+  //     case "image":
+  //     case "horizontal_rule": {
+  //       return EditRange.fromSelection(selection)
+  //     }
+  //   }
+  // }
+
+  // return markedRange(selection)
+}
+
+export const expandRange = (selection: Selection): ?NodeRange => {
+  const { $cursor, node, from, $from } = selection
+  const target = node || $from.doc.nodeAt($from.pos)
+  const blockRange =
+    $cursor != null
+      ? $cursor.blockRange()
+      : node != null ? selection.$from.blockRange(selection.$to) : null
+
+  if (blockRange) {
+    // const range = unionOfMarkers(getMarkersAt($cursor || selection.$from))
+    // const start = range.index - blockRange.start
+    // const end = blockRange.end - range.index - range.length
+    // range.
+  }
+
+  return blockRange
+}
+
+export const markedRange = (selection: Selection): EditRange =>
+  unionOfMarkers(getSelectionMarkers(selection), selection.$anchor.doc)
+
+export const unionOfMarkers = (markers: Marker[], doc: Node): EditRange => {
   let length = 0
   let index = Infinity
   for (let marker of markers) {
@@ -62,12 +182,12 @@ export const markedRange = (selection: Selection): Range => {
       length = Math.max(marker.end - index, length)
     }
   }
-  return new Range(index, length)
+  return EditRange.new(index, length, doc)
 }
 
 export const endEdit = (
   tr: Transaction,
-  range: Range,
+  range: EditRange,
   schema: Schema
 ): Transaction => {
   const { selection, doc } = tr
@@ -79,7 +199,7 @@ export const endEdit = (
 
 export const beginEdit = (
   tr: Transaction,
-  range: Range,
+  range: EditRange,
   schema: Schema
 ): Transaction => {
   const { selection, doc } = tr
@@ -139,7 +259,10 @@ class ChangeList {
   }
   insertMarkupCode(code: string, marks: Mark[]) {
     const { schema, meta, markup } = this
-    this.insertNode(schema.text(code, [meta, markup, ...marks]))
+    this.insertNode(schema.text(code, [meta, ...marks]))
+    // this.insertNode(
+    //   schema.node("Markup", { markup: code }, schema.text(code), marks)
+    // )
     return this
   }
   insertMarkup(markup: string, marks: Mark[]) {
@@ -183,35 +306,34 @@ export const expand = (
     const node = content.child(index)
     changeList.updateMarks(node.marks)
 
+    console.group(node.type.name)
     switch (node.type) {
       case schema.nodes.anchor: {
-        // changeList.index++
-        changeList.tr.setNodeMarkup(changeList.index, null, {
-          href: node.attrs.href,
-          title: node.attrs.title,
-          mode: "write"
-        })
         changeList.index++
-        // changeList.index++
-        // changeList.insertMarkupCode("[", node.marks)
-        // expand(node.content, changeList, schema)
-        // changeList.insertMarkupCode("](", node.marks)
+        changeList.insertMarkupCode("[", [])
+        expand(node.content, changeList, schema)
+        changeList.updateMarks([])
+        changeList.insertMarkupCode("]", [])
+        changeList.insertMarkupCode("(", [])
 
-        // const title =
-        // node.attrs.title == null
-        // ? ""
-        // : JSON.stringify(String(node.attrs.title))
+        const title =
+          node.attrs.title == null
+            ? ""
+            : JSON.stringify(String(node.attrs.title))
 
-        // changeList
-        // .insertMarkup(`${node.attrs.href} ${title}`, node.marks)
-        // .insertMarkupCode(")", node.marks)
-        // changeList.index++
+        changeList
+          .insertMarkup(`${node.attrs.href} ${title}`, node.marks)
+          .insertMarkupCode(")", node.marks)
+        changeList.index++
 
         break
       }
       case schema.nodes.heading: {
         const level: number = node.attrs.level || 1
-        changeList.markupNode(`${"#".repeat(level)} `, node)
+        changeList.index++
+        changeList.insertMarkupCode(`${"#".repeat(level)} `, node.marks)
+        // changeList.insertMarkup(" ", node.marks)
+        expand(node.content, changeList, schema)
         break
       }
       // case schema.nodes.horizontal_rule: {
@@ -240,17 +362,31 @@ export const expand = (
         changeList.retainNode(node)
         break
       }
-      default: {
+      case schema.nodes.paragraph: {
         changeList.index++
         expand(node.content, changeList, schema)
-        // changeList.retainNode(node)
+        break
+      }
+      default: {
+        // changeList.index++
+        // expand(node.content, changeList, schema)
+        changeList.retainNode(node)
         break
       }
     }
+    console.groupEnd()
     index++
   }
 
   return changeList
+}
+
+const isMarkup = (node: Node) =>
+  node.type.spec.markup || node.marks.some(isMarkupMark)
+
+const isMarkupMark = (mark: Mark) => {
+  const spec = mark.type.spec
+  return spec.markup || null ? true : false
 }
 
 export const collapse = (
@@ -263,15 +399,10 @@ export const collapse = (
   while (index < count) {
     const node = content.child(index)
     const marks = node.marks
-    if (marks.some($ => $.type.name === "markup")) {
+    if (node.type.spec.markup) {
       changeList.deleteNode(node)
-    } else if (node.type.name === "anchor") {
-      changeList.tr.setNodeMarkup(changeList.index, null, {
-        href: node.attrs.href,
-        title: node.attrs.title,
-        mode: "read"
-      })
-      changeList.index++
+    } else if (marks.some(isMarkupMark)) {
+      changeList.deleteNode(node)
     } else {
       if (node.isText) {
         changeList.retainNode(node)
@@ -287,32 +418,38 @@ export const collapse = (
   return changeList
 }
 
-export const editOffRange = (
+export const commitEdit = (
+  range: EditRange,
   tr: Transaction,
-  selection: Selection,
   schema: Schema
 ): ?Transaction => {
-  const { $cursor } = selection
-  if ($cursor) {
-    const { start, end } = $cursor.blockRange()
-    const slice = tr.doc.slice(start, end)
-    const source = slice.content.firstChild && slice.content.firstChild.content
-    if (!source || source.size === 0) {
-      return null
-    }
+  const delta = collapse(
+    tr.doc.slice(range.index, range.index + range.length).content,
+    new ChangeList(range.index, tr, schema, new Map()),
+    schema
+  )
+  const newRange = range.map(delta.tr.mapping, delta.tr.doc)
+  const slice = tr.doc.slice(newRange.index, newRange.index + newRange.length)
+  const markup = Serializer.serializeInline(slice.content)
+  const { content } = Parser.parse(`${markup}`)
 
-    const markup = Serializer.serializeInline(source)
-    const content = Parser.parseInline(markup)
+  const input = slice.content.textBetween(0, slice.content.size)
+  if (input !== content.textBetween(0, content.size)) {
+  }
 
-    const changeList = new ChangeList(start + 1, tr, schema, new Map())
-    return patch(source, content, changeList, schema)
+  const newTr = patch(
+    slice.content,
+    content,
+    new ChangeList(newRange.index, delta.tr, schema, new Map()),
+    schema
+  )
+  if (newTr) {
+    const range = newRange.map(newTr.mapping, newTr.doc)
+    return beginEdit(newTr, range, schema)
   } else {
     return null
   }
 }
-window.editOffRange = editOffRange
-window.Serializer = Serializer
-window.Parser = Parser
 
 export const patch = (
   last: Fragment,
@@ -324,9 +461,14 @@ export const patch = (
   const nextCount = next.childCount
   let lastIndex = 0
   let nextIndex = 0
+  const stepCount = changeList.tr.steps.length
 
   while (lastIndex < lastCount) {
     const lastChild = last.child(lastIndex)
+    if (isMarkup(lastChild)) {
+      lastIndex++
+      continue
+    }
     const nextChild = nextIndex < nextCount ? next.child(nextIndex) : null
 
     if (nextChild != null && lastChild.eq(nextChild)) {
@@ -334,8 +476,16 @@ export const patch = (
       lastIndex++
       nextIndex++
     } else {
-      changeList.deleteNode(lastChild)
-      lastIndex++
+      if (nextChild && lastChild.type === nextChild.type) {
+        changeList.index++
+        patch(lastChild.content, nextChild.content, changeList, schema)
+        changeList.index++
+        lastIndex++
+        nextIndex++
+      } else {
+        changeList.deleteNode(lastChild)
+        lastIndex++
+      }
     }
   }
 
@@ -344,5 +494,5 @@ export const patch = (
     nextIndex++
   }
 
-  return changeList.tr
+  return changeList.tr.steps.length - stepCount > 0 ? changeList.tr : null
 }
